use pointers and new
also included cpp files, showing all the errors in the code

+++ lib/Animation.cpp
@@ -16,7 +16,7 @@ String Animation::getName() { return this->name; }
 int Animation::getNumStrips() { return this->num_strips; }
 
 // get specific strip dependencies
-int* Animation::getDependencies() { return this->strips; }
+short int* Animation::getDependencies() { return this->strips; }
 
 // get update rate
 unsigned long int Animation::getUpdateRate() { return this->updateRate; }
